use a single SourceBufferControllerMock definition
